// src/types/types.d.ts
export type CityProps = {
    id: string;
    created_at: string;
    updated_at: string;
    promoted_at: string;
    width: number;
    height: number;
    color: string;
    blur_hash: string;
    description: string;
    alt_description: string;
    urls: {
      raw: string;
      full: string;
      regular: string;
      small: string;
      thumb: string;
      small_s3: string;
    };
    links: {
      self: string;
      html: string;
      download: string;
      download_location: string;
    };
    likes: number;
    liked_by_user: boolean;
    current_user_collections: [];
    sponsorship: null;
    topic_submissions: {};
    user: {
      id: string;
      updated_at: string;
      username: string;
      name: string;
      first_name: string;
      last_name: string;
      twitter_username: string;
      portfolio_url: string;
      bio: string;
      location: string;
      links: {
        self: string;
        html: string;
        photos: string;
        likes: string;
        portfolio: string;
        following: string;
        followers: string;
      };
      profile_image: {
        small: string;
        medium: string;
        large: string;
      };
      instagram_username: string;
      total_collections: number;
      total_likes: number;
      total_photos: number;
      accepted_tos: boolean;
      for_hire: boolean;
      social: {
        instagram_username: string;
        portfolio_url: string;
        twitter_username: string;
        paypal_email: string;
      };
      exif: {
        make: string;
        model: string;
        exposure_time: string;
        aperture: string;
        focal_length: string;
        iso: number;
      };
      location: {
        title: string;
        name: string;
        city: string;
        country: string;
        position: {
          latitude: number;
          longitude: number;
        };
      };
      meta: {
        index: boolean;
      };
      tags: {
        type: string;
        title: string;
      }[];
      tags_preview: {
        type: string;
        title: string;
      }[];
      views: number;
      downloads: number;
      topics: [];
    };
  };
  
  export type WeatherDataProps = {
    name: string;
    sys: {
      country: string;
    };
    weather: {
      description: string;
      icon: string;
    }[];
    main: {
      temp: number;
      feels_like: number;
      humidity: number;
    };
    wind: {
      speed: number;
    };
  };